switch from service to plugin
