+++ etc/adminhtml/di.xml
@@ -5,6 +5,9 @@
     <type name="Magento\Catalog\Model\Category">
         <plugin name="algolia_algoliasearch_plugin_category_url" type="Algolia\AlgoliaSearch\Plugin\CategoryUrlPlugin"/>
     </type>
+    <type name="Magento\Catalog\Controller\Adminhtml\Category\Edit">
+        <plugin name="algolia_algoliasearch_plugin_set_admin_current_category" type="Algolia\AlgoliaSearch\Plugin\SetAdminCurrentCategory"/>
+    </type>
     <type name="Algolia\AlgoliaSearch\Block\Adminhtml\LandingPage\Renderer\UrlBuilder">
         <arguments>
             <argument name="frontendUrlBuilder" xsi:type="object">Magento\Framework\Url</argument>
